Added the calendar_home page, also will do the failed login page with ajax validation tomorrow
In the index page ideally should be if successfully logged in then
header(location: calendar_home.php) if not then to the
failed_login.php. I'll add the ajax validation for the sign up form
tomorrow. Basic outline of the calendar day is added. And the login
form layout is fixed.